requirements.txt
requests
ts3
python-dotenv